# Conflicts:
#	docs/ChangeLog.md
#	docs/README.md
#	tests/common.mk
#	tests/runorder.mk

+++ tests/exit2.c
@@ -44,7 +44,7 @@
 static void *
 func(void * arg)
 {
-  int failed = (int) arg;
+  size_t failed = (size_t) arg;
 
   pthread_exit(arg);
 
@@ -52,7 +52,7 @@ func(void * arg)
   /*
    * Trick gcc compiler into not issuing a warning here
    */
-  assert(failed - (int)arg);
+  assert(failed - (size_t) arg);
 
   return NULL;
 }
