+++ tests/exit3.c
@@ -42,7 +42,7 @@
 static void *
 func(void * arg)
 {
-	int failed = (int) arg;
+	size_t failed = (size_t) arg;
 
 	pthread_exit(arg);
 
@@ -50,7 +50,7 @@ func(void * arg)
         /*
          * assert(0) in a way to prevent warning or optimising away.
          */
-	assert(failed - (int) arg);
+	assert(failed - (size_t) arg);
 
 	return NULL;
 }
